requirements.txt
openai
python-dotenv
semantic-kernel==0.3.13.dev